Added custom loot list presets
- Also added SOG PF Loot List
