+++ Headers/descriptionEXT/Loot Lists/Define Loot Lists.hpp
@@ -1,0 +1,8 @@
+#define LOOT_LIST_VALUES \
+    0,\
+    1
+
+// DO NOT CHANGE "OFF" POSITION
+#define LOOT_LIST_STRINGS \
+    "OFF", \
+    "SOG PF Loot List"
