+++ Headers/descriptionEXT/Loot Lists/Main Loot List.hpp
@@ -1,0 +1,71 @@
+class BLWK_lootLists
+{
+    class MasterLootList
+    {
+        lootBlackList[] = {
+            "O_Static_Designator_02_weapon_F",
+            "O_UAV_06_backpack_F",
+            "O_UAV_06_medical_backpack_F",
+            "O_UAV_01_backpack_F",
+            //"B_IR_Grenade",
+            "O_IR_Grenade",
+            "I_IR_Grenade",
+            "TrainingMine_Mag",
+            "ChemicalDetector_01_watch_F",
+            "ChemicalDetector_01_olive_F",
+            "8Rnd_82mm_Mo_Flare_white",
+            "muzzle_antenna_01_f",
+            "V_Safety_yellow_F",
+            "muzzle_antenna_02_f",
+            "rhsusf_100Rnd_556x45_M200_soft_pouch_ucp",
+            "rhsusf_100Rnd_556x45_M200_soft_pouch_coyote",
+            "rhsusf_100Rnd_556x45_M200_soft_pouch",
+            "rhs_mag_30Rnd_556x45_M200_Stanag",
+            "rhs_mag_20Rnd_556x45_M200_Stanag",
+            "rhsusf_100Rnd_762x51_m82_blank",
+            "rhsusf_50Rnd_762x51_m82_blank",
+            "hgun_Pistol_Signal_F"
+        };
+
+        lootWhitelist_launchers[] = {
+
+        };
+
+        lootWhitelist_primaries[] = {
+
+        };
+
+        lootWhitelist_handguns[] = {
+
+        };
+
+        lootWhitelist_backpacks[] = {
+
+        };
+
+        lootWhitelist_vests[] = {
+
+        };
+
+        lootWhitelist_uniforms[] = {
+
+        };
+
+        lootWhitelist_headgear[] = {
+
+        };
+
+        lootWhitelist_items[] = {
+
+        };
+
+        lootWhitelist_explosives[] = {
+
+        };
+    };
+
+    class CustomLootLists
+    {
+        #include "SOG PF Loot List.hpp"
+    };
+};
