+++ description.ext
@@ -110,4 +110,4 @@ class CfgNotifications
 
 #include "Headers\descriptionEXT\Build Items\Main Build Items.hpp"
 #include "Headers\descriptionEXT\Faction Headers\Faction Master.hpp"
-#include "Headers\descriptionEXT\Loot Lists.hpp"
+#include "Headers\descriptionEXT\Loot Lists\Main Loot List.hpp"
